+++ packages/infra/src/RequestContext.ts
@@ -1,4 +1,4 @@
-import { ExtendedTaggedClass, literal, NonEmptyString255, struct } from "@effect-app/schema"
+import { ExtendedTaggedClass, NonEmptyString255 } from "@effect-app/schema"
 import { S } from "effect-app"
 import { RequestId, UserProfileId } from "effect-app/ids"
 
@@ -8,8 +8,8 @@ export class RequestContextParent extends ExtendedTaggedClass<
 >()("RequestContext", {
   id: RequestId,
   name: NonEmptyString255,
-  userProfile: S.optional(struct({ sub: UserProfileId })),
-  locale: literal("en", "de"),
+  userProfile: S.optional(S.Struct({ sub: UserProfileId })),
+  locale: S.Literal("en", "de"),
   createdAt: S.Date.withDefault
 }) {}