+++ packages/infra/src/test.arbs.ts
@@ -1,10 +1,10 @@
 // Do not import to frontend
 
+import { FastCheck } from "@effect/schema"
 import { faker } from "@faker-js/faker"
 import { setFaker } from "effect-app/faker"
 import type { A } from "effect-app/schema"
 import { Random } from "fast-check"
-import * as fc from "fast-check"
 import * as rand from "pure-rand"
 
 const rnd = new Random(rand.congruential32(5))
@@ -14,14 +14,13 @@ setFaker(faker)
 /**
  * @tsplus getter FastCheck generateRandom
  */
-export function generate<T>(arb: fc.Arbitrary<T>) {
+export function generate<T>(arb: FastCheck.Arbitrary<T>) {
   return arb.generate(rnd, undefined)
 }
 
 /**
  * @tsplus getter effect/schema/Arbitrary generate
  */
-export function generateFromArbitrary<T>(arb: A.Arbitrary<T>) {
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-explicit-any
-  return generate(arb(fc as any))
+export function generateFromArbitrary<T>(arb: A.LazyArbitrary<T>) {
+  return generate(arb(FastCheck))
 }