+++ packages/schema/src/Class.ts
@@ -154,7 +154,7 @@ export const TaggedClass: <Self = never>(identifier?: string) => <Tag extends st
 ) => [Self] extends [never] ? MissingSelfGeneric<"Class">
   : EnhancedClass<
     Self,
-    { readonly "_tag": S.literal<[Tag]> } & Fields,
+    { readonly "_tag": S.Literal<[Tag]> } & Fields,
     Simplify<{ readonly _tag: Tag } & Struct.Type<Fields>>,
     Simplify<{ readonly _tag: Tag } & Struct.Encoded<Fields>>,
     Schema.Context<Fields[keyof Fields]>,
@@ -195,7 +195,7 @@ export const ExtendedTaggedClass: <Self, SelfFrom>(
   annotations?: S.Annotations.Schema<Self>
 ) => EnhancedClass<
   Self,
-  { readonly "_tag": S.literal<[Tag]> } & Fields,
+  { readonly "_tag": S.Literal<[Tag]> } & Fields,
   Simplify<{ readonly _tag: Tag } & Struct.Type<Fields>>,
   SelfFrom,
   Schema.Context<Fields[keyof Fields]>,