+++ packages/infra/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @effect-app/infra
 
+## 0.235.2
+
+### Patch Changes
+
+- Updated dependencies
+  - effect-app@0.204.0
+  - @effect-app/infra-adapters@0.155.2
+
+## 0.235.1
+
+### Patch Changes
+
+- ff5d466: cleanup Struct type
+- Updated dependencies [ff5d466]
+  - effect-app@0.203.1
+  - @effect-app/schema@0.221.1
+  - @effect-app/infra-adapters@0.155.1
+
+## 0.235.0
+
+### Minor Changes
+
+- 0d1b378: update to latest schema
+
+### Patch Changes
+
+- Updated dependencies [0d1b378]
+- Updated dependencies [256a4d1]
+  - @effect-app/infra-adapters@0.155.0
+  - effect-app@0.203.0
+  - @effect-app/schema@0.221.0
+  - @effect-app/core@0.170.0
+
 ## 0.234.7
 
 ### Patch Changes